requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.8.0
pillow>=9.0.0
pyyaml>=6.0
pydantic>=2.0.0
typing-extensions>=4.0.0

# Computer Vision & ML
mediapipe>=0.10.7
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Tracking
filterpy>=1.4.5
scipy>=1.9.0
scikit-learn>=1.1.0

# API & Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=11.0

# Logging & Monitoring
loguru>=0.7.0
tqdm>=4.64.0

# Data handling
pandas>=1.5.0
jsonlines>=3.1.0

# Testing
pytest>=7.2.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Development tools
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
pre-commit>=3.0.0

# Optional CUDA support (uncomment if GPU available)
# torch-audio>=2.0.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# torchvision>=0.15.0+cu118 --index-url https://download.pytorch.org/whl/cu118

# =============================================================================
# DEVELOPMENT DEPENDENCIES (Merged from requirements-dev.txt)
# =============================================================================

# Additional Testing
pytest-mock>=3.11.1

# Code formatting and linting
flake8>=6.0.0

# Type checking
types-PyYAML>=6.0.12
types-requests>=2.31.0

# Development tools
jupyterlab>=4.0.0
ipython>=8.14.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.1.0
mkdocs-mermaid2-plugin>=1.1.0

# Profiling and debugging
line-profiler>=4.1.0
memory-profiler>=0.61.0
py-spy>=0.3.14

# API development
httpx>=0.24.1
pytest-httpx>=0.24.1

# Data visualization (for development)
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.15.0

# Notebook integration
nbformat>=5.9.0
nbconvert>=7.7.0
